move style injection to OverlayBase component
